version 1 : Two separate concerns here.
 Logging error codes and choosing how to react to then.

This is coincidental duplication. and True duplication.

The logging is true duplication because there will most likely be a reason to log errors the same way.

The polices for dealing with server errors could change for certain services or resources. The fact that it is always handled the same way was coincidental. Not a truly shared trait between the different http verbs.